Initial website creation
I was still really confused on how and when to commit. I did not
realize that I should have been committing as I was creating each
page of the project. I went back after creating all of the pages
to learn how to stage, and commit to github because that has been the
most difficult part.

Project being uploaded, in full, now.